Player changed to PlayerProp to avoid common naming collisions.

+++ amara/amara_random.cpp
@@ -30,7 +30,7 @@ namespace Amara {
                 return dist(e);
             }
 
-            float random() {
+            double random() {
                 return between(0, 1);
             }
     };
